- fixed build for SLE
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@905 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ rpm/kiwi.spec
@@ -147,8 +147,9 @@ export K_USER=0 # set value to -1 to prevent building boot images
 rm -rf $RPM_BUILD_ROOT
 test -e /.buildenv || export K_USER=-1 # no buildenv, no boot image build
 test -e /.buildenv && . /.buildenv
-rm modules/KIWIScheme.rng
-#cat /proc/mounts > /etc/fstab
+if [ -x /usr/bin/java ];then
+	rm modules/KIWIScheme.rng
+fi
 make buildroot=$RPM_BUILD_ROOT CFLAGS="$RPM_OPT_FLAGS"
 
 if [ "$UID" = "$K_USER" ];then
